Get the killing of threads actually working.
