+++ dynamo_consistency/config.py
@@ -40,7 +40,7 @@ CONFIG = None
 SITE = None
 """
 A global place that stores a site that has been picked.
-Set in :py:func:`dynamo_consistency.summary.pick_site`.
+Set in :py:func:`dynamo_consistency.picker.pick_site`.
 """
 
 def config_dict():
